+++ packages/provisioner/README.md
@@ -1,2 +1,2 @@
-# truffle-provisioner
+# @truffle/provisioner
 Provision contract objects for use from multiple sources
