+++ packages/provisioner/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@truffle/provisioner",
+  "version": "0.1.5",
+  "description": "Provision contract objects for use from multiple sources",
+  "main": "index.js",
+  "scripts": {},
+  "repository": "https://github.com/trufflesuite/truffle/tree/master/packages/provisioner",
+  "keywords": [
+    "ethereum",
+    "truffle",
+    "contracts",
+    "provision"
+  ],
+  "author": "Tim Coulter <tim@trufflesuite.com>",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/trufflesuite/truffle/issues"
+  },
+  "homepage": "https://github.com/trufflesuite/truffle/tree/master/packages/provisioner#readme",
+  "publishConfig": {
+    "access": "public"
+  }
+}
