+++ packages/truffle-core/test/config.js
@@ -4,7 +4,7 @@ var glob = require("glob");
 var Box = require("@truffle/box");
 var Contracts = require("@truffle/workflow-compile");
 var Ganache = require("ganache-core");
-var provision = require("truffle-provisioner");
+var provision = require("@truffle/provisioner");
 var Resolver = require("truffle-resolver");
 var Artifactor = require("truffle-artifactor");
