--- packages/truffle-provisioner/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "truffle-provisioner",
-  "version": "0.1.5",
-  "description": "Provision contract objects for use from multiple sources",
-  "main": "index.js",
-  "scripts": {},
-  "repository": "https://github.com/trufflesuite/truffle/tree/master/packages/truffle-provisioner",
-  "keywords": [
-    "ethereum",
-    "truffle",
-    "contracts",
-    "provision"
-  ],
-  "author": "Tim Coulter <tim.coulter@consensys.net>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/trufflesuite/truffle-provisioner/issues"
-  },
-  "homepage": "https://github.com/trufflesuite/truffle-provisioner#readme",
-  "publishConfig": {
-    "access": "public"
-  }
-}
