+++ packages/truffle-resolver/index.js
@@ -5,7 +5,7 @@ var FSSource = require("./fs");
 var whilst = require("async/whilst");
 var contract = require("truffle-contract");
 var expect = require("@truffle/expect");
-var provision = require("truffle-provisioner");
+var provision = require("@truffle/provisioner");
 
 function Resolver(options) {
   expect.options(options, ["working_directory", "contracts_build_directory"]);
