+++ packages/truffle-resolver/package.json
@@ -24,7 +24,7 @@
     "detect-installed": "^2.0.4",
     "get-installed-path": "^4.0.8",
     "@truffle/expect": "^0.0.9",
-    "truffle-provisioner": "^0.1.5"
+    "@truffle/provisioner": "^0.1.5"
   },
   "devDependencies": {
     "mocha": "5.2.0",
